requirements.txt.txt
flask
flask-cors
scikit-learn
numpy
google-generativeai
openai
requests
gunicorn
